gvgai-assignment1-python/requirements.txt
imageio==2.31.2
pillow
pygame
